requirements.txt
click==5.1
Django==1.8.4
django-dotenv==1.3.0
django-solo==1.1.0
djangorestframework==3.3.1
wheel==0.24.0
